requirements.txt
openai>=1.37.0
tiktoken>=0.7.0
requests>=2.31.0
python-dotenv>=1.0.1
